Triggers. DailyBooleanMetricRest done. Next test Part 7

+++ src/main/java/ua/com/serverhelp/simplemetricstoragefile/entities/triggers/TimeElapsedFromLastEventExpression.java
@@ -15,7 +15,7 @@ import java.util.List;
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
-public class TimeElapsedFromLastEventExpression implements Expression<Long>{
+public class TimeElapsedFromLastEventExpression implements Expression<Double>{
     private Expression<List<DataElement>> arg1;
 
     @Override
@@ -32,7 +32,7 @@ public class TimeElapsedFromLastEventExpression implements Expression<Long>{
     }
 
     @Override
-    public Long getValue() throws ExpressionException {
+    public Double getValue() throws ExpressionException {
         List<DataElement> dataElements=arg1.getValue();
 
         if(dataElements.isEmpty()) throw new ExpressionException("Metric data is empty",new Exception());
@@ -41,7 +41,7 @@ public class TimeElapsedFromLastEventExpression implements Expression<Long>{
         Instant dataTime=Instant.ofEpochSecond(dataElement.getTimestamp());
         Duration duration=Duration.between(dataTime,Instant.now());
 
-        return duration.getSeconds();
+        return (double)duration.getSeconds();
     }
 
     @Override
